+++ src/index.ts
@@ -1,0 +1,7 @@
+export {GtExpandedRow} from './generic-table/components/gt-expanding-row/gt-expanding-row.component';
+export {GenericTableComponent} from './generic-table/generic-table.component';
+export {GtConfig} from './generic-table/interfaces/gt-config';
+export {GtInformation} from './generic-table/interfaces/gt-information';
+export {GtPagingInfo} from './generic-table/interfaces/gt-paging-info';
+export {GtTexts} from './generic-table/interfaces/gt-texts';
+export * from './generic-table/generic-table.module';